[feature] add hour enable verify after select hour and before confirm shift
